# openapi.yaml
openapi: 3.1.0
info:
  title: Sistema Integral de Gestión de Consorcios
  description: |
    API REST para la gestión integral de consorcios. Permite administrar expensas,
    gastos del consorcio, personal y liquidaciones, tareas y presupuestos,
    comunicación interna, reserva de amenities y configuración global. El acceso a
    los recursos está restringido por roles: Administración, Departamentos
    (Inquilinos/Propietarios), Representantes y Super-Admin.
    Fase 3.5 (jun-2026): cuenta corriente por departamento — pagos se asignan FIFO desde movimientos.

    **Multitenant SaaS (Fase 7A, jul-2026)**: la plataforma es multi-tenant con
    jerarquía Super-Admin → Administración → Consorcio → Departamento. Todo
    endpoint operacional (fuera de `/auth/*` y `/me/*`) exige el header
    `X-Consorcio-Id: <id>` para resolver el consorcio activo. El backend valida
    que el usuario del JWT tenga acceso al consorcio pedido; caso contrario
    devuelve **403**. Códigos específicos: `403 cambio_password_requerido`
    (usuario con `must_change_password=True`), `403 administracion_suspendida`
    (login sobre administración inactiva), `403 sin acceso a este consorcio`
    (spoof cross-tenant), `400 X-Consorcio-Id requerido/invalido`.
  version: 1.1.0
servers:
  - url: http://localhost:8000
    description: Servidor local de desarrollo
...
security:
  - bearerAuth: []
paths:
...
  /amenities/{amenity_id}/reservas:
    post:
      tags: [Amenities]
      summary: Reservar un amenity
      description: |
        Crea una nueva reserva para el amenity indicado. Accesible para todos
        los roles autenticados. El `usuario_id` se asigna server-side desde el
        token JWT; cualquier valor en el body es ignorado. `inicio` debe ser
        estrictamente anterior a `fin` (400 si no). Si el rango solicitado se
        superpone con una reserva confirmada existente para el mismo amenity,
        se devuelve 409 Conflict.
      operationId: crearReserva
      parameters:
        - name: amenity_id
          in: path
          required: true
          description: Identificador único del amenity.
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReservaCrear'
      responses:
        '201':
          description: Reserva creada exitosamente.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reserva'
        '400':
          $ref: '#/components/responses/PedidoInvalido'
        '401':
          $ref: '#/components/responses/NoAutenticado'
        '403':
          $ref: '#/components/responses/AccesoDenegado'
        '404':
          $ref: '#/components/responses/NoEncontrado'
        '409':
          description: El horario solicitado se superpone con una reserva existente.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /reservas:
    get:
      tags: [Amenities]
      summary: Listar mis reservas
      description: |
        Devuelve las reservas asociadas al usuario autenticado. Administración y
        Representantes ven todas las reservas. Los Departamentos ven solo las
        reservas que ellos mismos crearon. La identidad se obtiene del token.
      operationId: listarReservas
      parameters:
        - name: estado
          in: query
          required: false
          description: Filtrar por estado de la reserva.
          schema:
            type: string
            enum: [confirmada, cancelada]
      responses:
        '200':
          description: Lista de reservas recuperada exitosamente.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Reserva'
        '401':
          $ref: '#/components/responses/NoAutenticado'

  # =========================================================================
  # 5. ADMINISTRACIÓN (ALTAS Y EDICIÓN)
  # =========================================================================
...
  /reservas/{reserva_id}:
    get:
      tags: [Amenities]
      summary: Obtener detalle de una reserva
      operationId: obtenerReserva
      parameters:
        - name: reserva_id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200':
          description: Detalle de la reserva.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reserva'
        '401':
          $ref: '#/components/responses/NoAutenticado'
        '403':
          $ref: '#/components/responses/AccesoDenegado'
        '404':
          $ref: '#/components/responses/NoEncontrado'

    delete:
      tags: [Amenities]
      summary: Cancelar una reserva
      description: |
        Marca la reserva como `cancelada`. Solo el dueño de la reserva o
        Administración pueden cancelarla. Una reserva ya cancelada devuelve
        409 Conflict.
      operationId: cancelarReserva
      parameters:
        - name: reserva_id
          in: path
          required: true
          description: Identificador único de la reserva.
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Reserva cancelada exitosamente.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reserva'
        '401':
          $ref: '#/components/responses/NoAutenticado'
        '403':
          $ref: '#/components/responses/AccesoDenegado'
        '404':
          $ref: '#/components/responses/NoEncontrado'
        '409':
          description: La reserva ya está cancelada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /proveedores:
    get:
      tags: [Configuración]
      summary: Listar proveedores
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: activo
          required: false
          schema: { type: boolean, default: true }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ProveedorOut' }
        '401': { $ref: '#/components/responses/NoAutenticado' }
        '403': { $ref: '#/components/responses/AccesoDenegado' }
    post:
      tags: [Configuración]
      summary: Crear proveedor
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProveedorCrear' }
      responses:
        '201':
          description: Creado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProveedorOut' }
        '400': { $ref: '#/components/responses/PedidoInvalido' }
        '401': { $ref: '#/components/responses/NoAutenticado' }
        '403': { $ref: '#/components/responses/AccesoDenegado' }
        '409':
          description: Ya existe un proveedor con ese CUIT.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /proveedores/{proveedor_id}:
    parameters:
      - in: path
        name: proveedor_id
        required: true
        schema: { type: integer }
    get:
      tags: [Configuración]
      summary: Obtener proveedor
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProveedorOut' }
        '401': { $ref: '#/components/responses/NoAutenticado' }
        '403': { $ref: '#/components/responses/AccesoDenegado' }
        '404': { $ref: '#/components/responses/NoEncontrado' }
    patch:
      tags: [Configuración]
      summary: Editar proveedor
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProveedorActualizar' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProveedorOut' }
        '400': { $ref: '#/components/responses/PedidoInvalido' }
        '401': { $ref: '#/components/responses/NoAutenticado' }
        '403': { $ref: '#/components/responses/AccesoDenegado' }
        '404': { $ref: '#/components/responses/NoEncontrado' }
    delete:
      tags: [Configuración]
      summary: Desactivar proveedor (soft-delete)
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: OK (proveedor desactivado)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProveedorOut' }
        '401': { $ref: '#/components/responses/NoAutenticado' }
        '403': { $ref: '#/components/responses/AccesoDenegado' }
        '404': { $ref: '#/components/responses/NoEncontrado' }
...
  /reportes/proveedores:
    get:
      tags: [Reportes]
      summary: Lista de proveedores con totales facturados
      operationId: listarProveedoresReporte
      security: [{bearerAuth: []}]
      parameters:
        - name: anio
          in: query
          schema: { type: integer }
        - name: periodo
          in: query
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ItemProveedorOut' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Autenticación mediante token JWT en el header Authorization.
...
  responses:
    PedidoInvalido:
      description: El pedido es inválido o le faltan campos requeridos.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NoAutenticado:
      description: No se proporcionó token de autenticación o es inválido.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    AccesoDenegado:
      description: El rol del usuario no tiene permisos para acceder al recurso.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NoEncontrado:
      description: El recurso solicitado no existe.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      required: [detail]
      properties:
        detail:
          type: string
          description: Descripción legible del error.
          examples: ["No tiene permisos para acceder a este recurso."]

    # ---------------------------- Consorcios (Plan D) ---------------------
...
    Reserva:
      type: object
      required: [id, amenity_id, usuario_id, inicio, fin, estado, movimiento_cuenta_id]
      properties:
        id: { type: integer, format: int64 }
        amenity_id: { type: integer, format: int64 }
        usuario_id: { type: integer, format: int64 }
        inicio: { type: string, format: date-time }
        fin: { type: string, format: date-time }
        estado: { type: string, enum: [confirmada, cancelada] }
        movimiento_cuenta_id: { type: integer, format: int64, nullable: true }
    ReservaCrear:
      type: object
      required: [inicio, fin]
      properties:
        inicio:
          type: string
          format: date-time
        fin:
          type: string
          format: date-time
...
    ProveedorCrear:
      type: object
      required: [razon_social, cuit]
      properties:
        razon_social: { type: string, minLength: 1, maxLength: 255 }
        nombre_fantasia: { type: string, maxLength: 255, nullable: true }
        cuit: { type: string, pattern: '^\d{2}-\d{8}-\d{1}$' }
        direccion: { type: string, maxLength: 500, nullable: true }
    ProveedorActualizar:
      type: object
      properties:
        razon_social: { type: string, minLength: 1, maxLength: 255, nullable: true }
        nombre_fantasia: { type: string, maxLength: 255, nullable: true }
        direccion: { type: string, maxLength: 500, nullable: true }
        activo: { type: boolean, nullable: true }
    ProveedorOut:
      type: object
      required: [id, razon_social, cuit, activo]
      properties:
        id: { type: integer }
        razon_social: { type: string }
        nombre_fantasia: { type: string, nullable: true }
        cuit: { type: string }
        direccion: { type: string, nullable: true }
        activo: { type: boolean }
...
    ItemProveedorOut:
      type: object
      properties:
        proveedor_id: { type: integer }
        razon_social: { type: string }
        cuit: { type: string }
        cantidad_gastos: { type: integer }
        total_facturado: { type: number }
        ultimo_gasto: { type: string, format: date, nullable: true }